[ALPS08477052] cmdq: change reuse offset size
[Description]
change reuse offset size to 32bit
[Test]
oss8 Tuota/ ATE

MTK-Commit-Id: 23da58162d948a6ec30942e85f058019d0c30c24

CR-Id: ALPS08477052
Feature: cmdq_gce

+++ include/linux/soc/mediatek/mtk-cmdq-ext.h
@@ -323,7 +323,7 @@ struct cmdq_flush_completion {
 struct cmdq_reuse {
 	u64 *va;
 	u32 val;
-	u16 offset;
+	u32 offset;
 	u8 op;
 };
